Add the necessary files
js/script.js - Create a JavaScript file inside the "js" folder and link it to index.html

css/reset.css or css/normalize.css - This is not a project requirement, but using a reset or normalize CSS file is an important common standard in web development as it helps zero out the initial styles that different browsers add to their page by default. So you are encouraged to take this extra step. You should be able to find what you need with a Google search.