Add better validation
Now the credit card, zip and cvv fields are checked to see if they are integers. Also, If we choose the card and begin entering numbers and then stop midway, the error will appear. If we then choose another payment method, the card error is still there and we can’t submit, even though we chose another payment.